Serialize and deserialize 110a and 110b EML files
